graph: Add graph implementation
Added a simple graph implementation which consists of a vector of the
neighbours of each node. The edge weights can be customised using a
weight policy. One weight policy is implemented so far, is just gives
each edge a constant weight of one. The edge weights will be the
interaction strengths of the neighbouring spins.